replicon project task
developed the code using .Net to solve the following TWO problem statements

State Bank of India is planning to introduce Automated Teller Machine (ATM) system
for its customers to provide convenient withdrawal of Cash. A software is to be designed
that will control ATM dispenser for cash (in multiples of Rs.500, 100, 50, 10 and 5 only)
Write a program which will take Amount as input and gives output as break up of how
many notes of each denomination will be dispensed.
If not able to dispense amount, it should show error as &#39;Invalid Amount&#39;
For e.g.:
555 as input will return the output as:
500 * 1
50 * 1
5 * 1
Problem Statement 2:
Write a program to find out factorial of a number. Write using both recursion and
Iteration techniques.
For e.g.:
Input 5
Factorial is (5 * 4 * 3 * 2 * 1) so output 120
Input 6
Factorial is (6 * 5 * 4 * 3 * 2 * 1) so output 720